move fakeplayer uuid logic out of a random feature
